A lot of thinks

+++ src/main/java/org/example/dto/ApartmentDTO.java
@@ -4,13 +4,15 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.example.entity.Apartment;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class ApartmentDTO {
     private Long id;
     private int number;
-    private double area;
+    private BigDecimal area;
     public ApartmentDTO(Apartment apartment) {
         this.id = apartment.getId();
         this.number = apartment.getNumber();
